@@ -1,11 +1,12 @@
 {
-    "name": "wearejust/laravel-amp",
+    "name": "cohensive/laravel-amp",
     "description": "Package that helps you set up AMP (Accelerated Mobile Pages) using Laravel",
     "keywords": [
         "wearejust",
+        "cohensive",
         "laravel-amp"
     ],
-    "homepage": "https://github.com/wearejust/amp",
+    "homepage": "https://github.com/kanecohen/laravel-amp",
     "license": "MIT",
     "authors": [
         {
@@ -13,24 +14,28 @@
             "email": "cees.vanegmond@wearejust.com",
             "homepage": "https://wearejust.com",
             "role": "Developer"
+        },
+        {
+            "name": "Kane Cohen",
+            "email": "kanecohen@gmail.com",
+            "role": "Developer"
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "illuminate/support": "~5",
-        "illuminate/contracts": "~5"
+        "php": "^8.0",
+        "laravel/framework": "^5.8|^6.0|^7.0|^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "5.*"
+        "phpunit/phpunit": "^8.0|^9.0"
     },
     "autoload": {
         "psr-4": {
-            "Just\\Amp\\": "src"
+            "Cohensive\\Amp\\": "src"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Just\\Amp\\Test\\": "tests"
+            "Cohensive\\Amp\\Test\\": "tests"
         }
     },
     "scripts": {
@@ -42,10 +47,10 @@
     "extra": {
         "laravel": {
             "providers": [
-                "Just\\Amp\\Laravel\\AmpServiceProvider"
+                "Cohensive\\Amp\\AmpServiceProvider"
             ],
             "aliases": {
-                "AmpRouter": "Just\\Amp\\Laravel\\AmpRouteFacade"
+                "AmpRouter": "Cohensive\\Amp\\AmpRouteFacade"
             }
         }
     }
