@@ -1,11 +1,9 @@
 <?php
+use PHPUnit\Framework\TestCase;
 
-namespace Just\Amp\Test;
-
-class AmpRouterTest extends \PHPUnit_Framework_TestCase
+class AmpRouterTest extends TestCase
 {
-    /** @test */
-    public function true_is_true()
+    public function testTrueIsTrue()
     {
         $this->assertTrue(true);
     }
